for heroku postgresql

+++ conf/evolutions/default/1.sql
@@ -3,7 +3,7 @@
 # --- !Ups
 
 create table user (
-  id bigint not null primary key,
+  id integer not null primary key,
   email varchar(255) not null,
   name varchar(255) not null,
   password varchar(255) not null
